Allow characters like \( and \} to parse
The issue is the parser was assuming that what follows the backslash
is an atom (I presume like a symbol) but characters like parens should
be allowed here when they're not in atoms. Added some customized logic
for reading chars, not particularly optimized but I don't think it's a
big bottleneck.

Added some basic tests.